server/requirements.txt
alembic==1.0.7
atomicwrites==1.3.0
attrs==18.2.0
bcrypt==3.1.6
cffi==1.12.0
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.17.0
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-Redis==0.3.0
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==6.0.0
pluggy==0.8.1
py==1.7.0
pycparser==2.19
PyJWT==1.4.2
PyMySQL==0.9.3
pytest==4.2.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
redis==3.2.0
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
WTForms==2.2.1
